refactor(db-client-ts): finish grouping fsm-sync-worker-v1 ops, retire   fsm-instance.ts (#130)
  Continue the src/ reorg into numbered load-order folders:

  - 35_fsm_sync_operation_worker_v1/fsmSyncOperationScheduler.ts —
    scheduleNextPending (was fsm-workerlet.ts)
  - 35_fsm_sync_operation_worker_v1/fsmSyncOperationWorkerlet.ts —
    renamed+moved fsm-workerlet.ts (fsm_workerlet table ops)
  - 35_fsm_sync_operation_worker_v1/fsmctl.ts — renamed+moved fsm-scheduler.ts
    (enqueueDispatch/resumeEventForFsmWorker), then absorbed fsm-instance.ts's
    instance lifecycle (listFsmInstances, isFSMInstancePresent,
    createFsmInstanceFromName, stopEventForFsmWorker, getFSMData,
    getFsmDataResolveStateValue) and event ops (sendEventToFsmQueueWithEventLogs,
    archiveEventFromFsmTypeWorker); logger category renamed "scheduler" ->
    "fsmctl" to match its now-broader scope
  - 30_async_operation_worker_v2/asyncOperationCtl.ts — new file holding
    archiveEventFromFsmPromiseTypeWorker, fsm-instance.ts's last remaining
    export; fsm-instance.ts is now empty and deleted
  - 25_async_operation_worker_v1/asyncOperationCtl.ts renamed to
    asyncOperationWorkerCtl.ts (disambiguates from the new
    30_async_operation_worker_v2/asyncOperationCtl.ts)

  Update index.ts's barrel exports and CLAUDE.md/README.md file listings to
  match. All downstream consumers import through the @pgfsm/db package barrel,
  so no call-site changes were needed.
